Add Inject Resolving and restructure scoped and context dependency
